pulse: carry the server tick as f64, and keep one routing transport per realm
Two defects that between them froze foreign avatars, both found by
instrumenting a live session.

The server tick was laundered through `rfc4::Movement.timestamp`, a proto
`float`, on the way to the interpolator's strictly-newer gate. The tick is
absolute milliseconds (~2.4e9), where an `f32` has a 256ms quantum — so
consecutive deltas ~100ms apart collapsed onto one value, `forward` came out as
exactly 0.0, and the gate discarded them as stale duplicates. Captured ticks
2350530722 / 821 / 940 all read 2350530.75. The updates lost that way were the
ones decelerating a peer and then zeroing its velocity, leaving the avatar
dead-reckoning at its last-known speed until a tick gap happened to exceed one
ulp — roughly six stops in eight went unterminated.

Carry it as an `f64` on `PulseEvent::Movement` / `PlayerMessage::Movement`
instead, passed to `apply_foreign_movement` explicitly the way `teleport`
already is, and widen the event and target fields to match. Absolute
milliseconds are exact in `f64` for any tick the protocol can produce, so this
needs no epoch and no modulus, and the value stays a real monotonic time for
future interpolation work rather than becoming an ordering token. The LiveKit
arm keeps its own sender-clock value, widened losslessly.

`start_pulse` also spawned a routing transport for every `StartPulse`, which
fires for archipelago island hops as well as realm changes, while nothing
despawned the old one. Each hop therefore left another live routing entity, and
`broadcast` copies every message onto all of them: two islands doubled a 10Hz
send rate into the server's 20Hz `PlayerStateInput` cap and earned an
`InputRateExceeded` disconnect, which is classified terminal. Pulse then stayed
down for the rest of the session while LiveKit carried on, so peers kept their
animations and never moved again.

Only a realm change needs a new transport, so gate on the realm. An island hop
is now a no-op, and deliberately so: rebuilding would empty every Pulse peer's
`ForeignPlayer.transports` for a frame and — with the inactivity grace period
gone — despawn and respawn them, firing playerDisconnected/onLeaveScene for
people who never left. Surviving a hop is precisely what riding Pulse should buy
them, since unlike a livekit island their transport has not changed. On a real
realm change the sweep has already queued our entity for despawn, so drop the id
we own explicitly and spawn fresh rather than racing the deferred flush.

+++ crates/avatar/src/foreign_dynamics.rs
@@ -14,7 +14,7 @@ use scene_runner::{update_world::mesh_collider::SceneColliderData, ContainingSce
 
 /// A backward jump beyond this reads as a server restart (the Pulse `server_tick` resets to ~0), so
 /// we re-sync to it instead of freezing; smaller backward steps are reordered/duplicate datagrams.
-const TIMESTAMP_RESET_SECS: f32 = 60.0;
+const TIMESTAMP_RESET_SECS: f64 = 60.0;
 
 pub struct PlayerMovementPlugin;
 
@@ -35,7 +35,8 @@ impl Plugin for PlayerMovementPlugin {
 #[derive(Component)]
 struct PlayerTargetPosition {
     time: f32,
-    timestamp: Option<f32>,
+    /// The source clock of the last applied update, seconds (see `PlayerPositionEvent`).
+    timestamp: Option<f64>,
     velocity: Option<Vec3>,
     /// Already box-resolved against the quantization grid when the packet landed (see
     /// `update_foreign_user_target_position`); the interpolator treats it as an exact, stable
